Added - Authentication Change - All controller        - check condition before remove        - Added booking table

+++ src/main/java/com/group4/carrental/controllers/UserListedCarController.java
@@ -1,5 +1,7 @@
 package com.group4.carrental.controllers;
 
+import com.group4.carrental.authentication.Authentication;
+import com.group4.carrental.authentication.IAuthentication;
 import com.group4.carrental.model.CarList;
 import com.group4.carrental.model.CarType;
 import com.group4.carrental.model.City;
@@ -14,6 +16,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.servlet.http.HttpSession;
+import javax.swing.undo.AbstractUndoableEdit;
 import java.util.ArrayList;
 
 @Controller
@@ -21,6 +24,7 @@ public class UserListedCarController {
 
     private IUserListedCarService userListedCarService;
     private LoggerInstance loggerInstance;
+    private Authentication authentication = Authentication.getInstance();
 
     @Autowired
     public UserListedCarController(@Qualifier("UserListedCarService") IUserListedCarService userListedCarService,
@@ -31,64 +35,62 @@ public class UserListedCarController {
 
 
     @GetMapping("/user-listed-cars")
-    public String userListedCars(Model model, HttpSession session){
-        loggerInstance.log(0,"User Listed Cars: Called");
-        int userId = 0;
+    public String userListedCars(Model model, HttpSession session) {
+        loggerInstance.log(0, "User Listed Cars: Called");
 
-        try {
-            userId = (int) session.getAttribute("user_id");
-        }catch (NullPointerException exception){
+        if(authentication.isValidUserSession(session)){
+            int userId = authentication.getUserId();
+            ArrayList<CarList> listedCars = userListedCarService.getUserListedCars(userId);
+            model.addAttribute("listedCars",listedCars);
+            return  "userListedcar";
+        }else {
             return "redirect:login";
         }
 
-
-
-        ArrayList<CarList> listedCars = userListedCarService.getUserListedCars(userId);
-
-        model.addAttribute("listedCars",listedCars);
-        return  "userListedcar";
 
     }
 
     @PostMapping("/user-listed-cars")
     public String removeCarFromListedCar(@RequestParam("carId") int carId, Model model, HttpSession session){
         loggerInstance.log(0,"User Remove Car From Listed Car: Called");
-        int userId = 0;
-        try {
-            userId = (int) session.getAttribute("user_id");
-        }catch (NullPointerException exception){
+        if(authentication.isValidUserSession(session)){
+            int userId = authentication.getUserId();
+            boolean isBooked = userListedCarService.isCarBooked(carId);
+            if(isBooked){
+                model.addAttribute("error","You can not remove Booked Car");
+            }else {
+                userListedCarService.removeCarById(carId);
+            }
+
+            ArrayList<CarList> listedCars = userListedCarService.getUserListedCars(userId);
+
+            model.addAttribute("listedCars",listedCars);
+            return "userListedcar";
+        }else {
             return "redirect:login";
         }
 
-        System.out.println("car id = "+carId);
-        userListedCarService.removeCarById(carId);
-        ArrayList<CarList> listedCars = userListedCarService.getUserListedCars(userId);
-
-        model.addAttribute("listedCars",listedCars);
-        return "userListedcar";
     }
 
     @GetMapping("edit-car-details")
     public String editCarDetails(@RequestParam("carIdEdit") int carId,HttpSession session, Model model){
         loggerInstance.log(0,"User Car Edit Details: Called");
-        int userId = 0;
-        try {
-            userId = (int) session.getAttribute("user_id");
-        }catch (NullPointerException exception){
+        if(authentication.isValidUserSession(session)){
+            CarList carDetails = userListedCarService.getCarDetailsById(carId);
+
+            ArrayList<CarType> carTypeArrayList = userListedCarService.getCarTypeList();
+            model.addAttribute("carType",carTypeArrayList);
+            ArrayList<City> cityArrayList = userListedCarService.getCityList();
+            model.addAttribute("city",cityArrayList);
+
+
+            model.addAttribute("carDetails",carDetails);
+
+
+            return "carEdit";
+        }else {
             return "redirect:login";
         }
 
-        CarList carDetails = userListedCarService.getCarDetailsById(carId);
-
-        ArrayList<CarType> carTypeArrayList = userListedCarService.getCarTypeList();
-        model.addAttribute("carType",carTypeArrayList);
-        ArrayList<City> cityArrayList = userListedCarService.getCityList();
-        model.addAttribute("city",cityArrayList);
-
-
-        model.addAttribute("carDetails",carDetails);
-
-
-        return "carEdit";
     }
 }
